social_media_api/requirements.txt
Django>=5.0.0
djangorestframework>=3.14.0
Pillow>=10.0.0
